AI-Agentic/google-trends/requirements.txt
playwright
pandas
python-dotenv
bs4
google-genai